函数和 lambda

+++ src/redscript/compiler/TokenSpace.scala
@@ -1,16 +1,18 @@
 package redscript.compiler
 
+import java.math.BigInteger
+
 import scala.util.parsing.combinator.token.StdTokens
 
 trait TokenSpace extends StdTokens
 {
-    case class IntLit(value: Long) extends Token
+    case class IntLit(value: Int) extends Token
     {
         override def chars: String = value.toString
         override def toString: String = chars
     }
 
-    case class LongLit(value: BigInt) extends Token
+    case class LongLit(value: BigInteger) extends Token
     {
         override def chars: String = value.toString
         override def toString: String = chars
